# conecteai/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Conecte.ai",
        "version": "1.0"
    },
    "paths": {
...
        "/api/order": {
            "get": {
                "tags": [
                    "OrderController"
                ],
                "summary": "Lista os pedidos cadastrados.",
                "description": "Lista os pedidos cadastrados no banco de dados.",
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "example": [
                                    "$order"
                                ]
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "OrderController"
                ],
                "summary": "Cria um pedido.",
                "description": "Cria um novo pedido no banco de dados.",
                "parameters": [
                    {
                        "name": "salesman",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Felipe Kazuo"
                    },
                    {
                        "name": "customer_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1"
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "03/03/2021"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Pendente"
                    },
                    {
                        "name": "product_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1"
                    },
                    {
                        "name": "total_price",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "100"
                    },
                    {
                        "name": "commission",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "2.5"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "example": [
                                    "Erro ao cadastrar pedido."
                                ]
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "example": [
                                    "Cliente não encontrado."
                                ]
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "example": [
                                    "Erro ao criar pedido."
                                ]
                            }
                        }
                    },
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "example": [
                                    "Pedido efetuado com sucesso."
                                ]
                            }
                        }
                    }
                }
            }
        },
        "/api/order/{id}": {
            "get": {
                "tags": [
                    "OrderController"
                ],
                "summary": "Mostra um pedido cadastrado.",
                "description": "Mostra um pedido cadastrado no banco de dados.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "example": [
                                    "Pedido não encontrado."
                                ]
                            }
                        }
                    },
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "example": [
                                    "$id"
                                ]
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "OrderController"
                ],
                "summary": "Atualiza um pedido.",
                "description": "Atualiza um pedido existente no banco de dados.",
                "parameters": [
                    {
                        "name": "salesman",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Felipe Kazuo"
                    },
                    {
                        "name": "cusomer_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1"
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "03/03/2021"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Pago"
                    },
                    {
                        "name": "product_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1"
                    },
                    {
                        "name": "total_price",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "100"
                    },
                    {
                        "name": "commission",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "2.5"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "example": [
                                    "Erro ao atualizar pedido."
                                ]
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "example": [
                                    "Cliente não encontrado."
                                ]
                            }
                        }
                    },
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "example": [
                                    "Pedido atualizado com sucesso."
                                ]
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "OrderController"
                ],
                "summary": "Deleta um pedido.",
                "description": "Deleta um pedido existente no banco de dados.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "example": [
                                    "Erro ao deletar pedido."
                                ]
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "example": [
                                    "Pedido não encontrado."
                                ]
                            }
                        }
                    },
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "example": [
                                    "Pedido deletado com sucesso."
                                ]
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Faça login com email e senha para obter seu token.",
                "name": "Token",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}